requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pytest==8.0.2
pytest-django==4.8.0
coverage==7.4.1
drf-nested-routers==0.93.5
django-filter==23.5
drf-spectacular==0.27.1 